// frontend/src/api/generated/models/productionCostResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface ProductionCostResponse {
  id: string;
  production_order_id: string;
  order_number: string;
  component: string;
  source_type: string;
  source_id: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  amount: string;
  allocation_basis: string;
}